Character extraction
a) Create a c-string that contains: "5,3,B,1,8,c,5,E,6". 
b) Create a function sumDigits that returns the sum of all numbers within the string. Note that this will require type conversion (casting). 
c) Create a function sumChars that returns a string object that contains the concatenation of all chars within the original c-string.  
d) Output the integer sum and concatenated string to console.
 
Output Example 
Sum of digits: 28 
Sum of chars: BcE