Recursive function 01
Write a program that  prompts an receives two integers between 20 and 40 and call a recursive function that finds the product of all numbers within the range exclusively.

a) Main should handle all input and output.   
b) Create a function that accepts a two numbers and finds the product of the range using a recursive process to do so.    c) Demonstrate the function with sample input from console.

For example, 20 and 24; the product =20*21*22*23*24
 
Example Output (input in bold italics)
Enter the minimum and maximum of the range (between 20 and 40) > 20 24
The product of the range of 20 and 24 = 5100480